# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Asset Care API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/sync/pull": {
      "get": {
        "tags": [
          "Sync"
        ],
        "summary": "Sync Pull",
        "description": "Deliver all assets, rules, and logs modified or deleted since the provided timestamp.\n\nIntent:\n    Enables incremental delta downloads. Soft-deleted records are included\n    so the mobile SQLite database can apply local deletions.",
        "operationId": "sync_pull_api_v1_sync_pull_get",
        "parameters": [
          {
            "name": "since",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "description": "UTC timestamp of the client's last successful synchronization",
              "title": "Since"
            },
            "description": "UTC timestamp of the client's last successful synchronization"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncPullResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sync/push": {
      "post": {
        "tags": [
          "Sync"
        ],
        "summary": "Sync Push",
        "description": "Reconcile offline-created or modified mutations using Last-Write-Wins strategy.\n\nIntent:\n    Processes mutations submitted by the client while disconnected.\n    Uses client-generated UUIDs to ensure seamless identity persistence across SQLite and PostgreSQL.",
        "operationId": "sync_push_api_v1_sync_push_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncPushRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncPushResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssetResponse": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "description": "Asset display name"
          },
          "asset_type": {
            "$ref": "#/components/schemas/AssetType",
            "description": "Classification of the asset"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Optional notes or details"
          },
          "metadata_payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata Payload",
            "description": "Category-specific dynamic attributes"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "is_deleted": {
            "type": "boolean",
            "title": "Is Deleted"
          }
        },
        "type": "object",
        "required": [
          "name",
          "asset_type",
          "id",
          "created_at",
          "updated_at",
          "is_deleted"
        ],
        "title": "AssetResponse",
        "description": "Response DTO for an asset."
      },
      "AssetType": {
        "type": "string",
        "enum": [
          "vehicle",
          "hvac",
          "appliance",
          "other"
        ],
        "title": "AssetType",
        "description": "Categorization of supported asset classes."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MaintenanceLogResponse": {
        "properties": {
          "maintenance_type": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Maintenance Type",
            "description": "e.g. oil_change, filter_swap"
          },
          "service_date": {
            "type": "string",
            "format": "date-time",
            "title": "Service Date",
            "description": "UTC timestamp or ISO date of service execution"
          },
          "metric_value_at_service": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Metric Value At Service",
            "description": "Odometer or usage hours at service time"
          },
          "cost": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,2}0*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost",
            "description": "Service cost"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes",
            "description": "Observations, parts replaced"
          },
          "performed_by": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Performed By",
            "description": "Technician or workshop name"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "asset_id": {
            "type": "string",
            "format": "uuid",
            "title": "Asset Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "is_deleted": {
            "type": "boolean",
            "title": "Is Deleted"
          }
        },
        "type": "object",
        "required": [
          "maintenance_type",
          "service_date",
          "id",
          "asset_id",
          "created_at",
          "updated_at",
          "is_deleted"
        ],
        "title": "MaintenanceLogResponse",
        "description": "Response DTO for an executed maintenance log."
      },
...
      "MaintenanceRuleResponse": {
        "properties": {
          "maintenance_type": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Maintenance Type",
            "description": "e.g. oil_change, air_filter"
          },
          "metric_type": {
            "$ref": "#/components/schemas/MetricType",
            "description": "Trigger metric: odometer, time_days, time_months, usage_hours"
          },
          "interval_value": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Interval Value",
            "description": "Numerical interval (e.g. 5000, 180, 6)"
          },
          "metric_unit": {
            "type": "string",
            "maxLength": 20,
            "minLength": 1,
            "title": "Metric Unit",
            "description": "Measurement unit (e.g. km, mi, months)"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Optional service notes"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "asset_id": {
            "type": "string",
            "format": "uuid",
            "title": "Asset Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "is_deleted": {
            "type": "boolean",
            "title": "Is Deleted"
          }
        },
        "type": "object",
        "required": [
          "maintenance_type",
          "metric_type",
          "interval_value",
          "metric_unit",
          "id",
          "asset_id",
          "created_at",
          "updated_at",
          "is_deleted"
        ],
        "title": "MaintenanceRuleResponse",
        "description": "Response DTO for a maintenance rule."
      },
...
      "MetricType": {
        "type": "string",
        "enum": [
          "odometer",
          "time_days",
          "time_months",
          "usage_hours"
        ],
        "title": "MetricType",
        "description": "Types of metrics used to trigger maintenance thresholds."
      },
      "SyncAssetItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "asset_type": {
            "$ref": "#/components/schemas/AssetType"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "metadata_payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata Payload"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "is_deleted": {
            "type": "boolean",
            "title": "Is Deleted",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "asset_type",
          "updated_at"
        ],
        "title": "SyncAssetItem",
        "description": "Asset mutation item submitted during push."
      },
      "SyncLogItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "asset_id": {
            "type": "string",
            "format": "uuid",
            "title": "Asset Id"
          },
          "maintenance_type": {
            "type": "string",
            "title": "Maintenance Type"
          },
          "service_date": {
            "type": "string",
            "format": "date-time",
            "title": "Service Date"
          },
          "metric_value_at_service": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metric Value At Service"
          },
          "cost": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "performed_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Performed By"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "is_deleted": {
            "type": "boolean",
            "title": "Is Deleted",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "asset_id",
          "maintenance_type",
          "service_date",
          "updated_at"
        ],
        "title": "SyncLogItem",
        "description": "Maintenance log mutation item submitted during push."
      },
      "SyncPullResponse": {
        "properties": {
          "synced_at": {
            "type": "string",
            "format": "date-time",
            "title": "Synced At"
          },
          "assets": {
            "items": {
              "$ref": "#/components/schemas/AssetResponse"
            },
            "type": "array",
            "title": "Assets"
          },
          "maintenance_rules": {
            "items": {
              "$ref": "#/components/schemas/MaintenanceRuleResponse"
            },
            "type": "array",
            "title": "Maintenance Rules"
          },
          "maintenance_logs": {
            "items": {
              "$ref": "#/components/schemas/MaintenanceLogResponse"
            },
            "type": "array",
            "title": "Maintenance Logs"
          }
        },
        "type": "object",
        "required": [
          "synced_at",
          "assets",
          "maintenance_rules",
          "maintenance_logs"
        ],
        "title": "SyncPullResponse",
        "description": "Payload sent to mobile client containing all entities modified since request timestamp."
      },
      "SyncPushRequest": {
        "properties": {
          "assets": {
            "items": {
              "$ref": "#/components/schemas/SyncAssetItem"
            },
            "type": "array",
            "title": "Assets"
          },
          "maintenance_rules": {
            "items": {
              "$ref": "#/components/schemas/SyncRuleItem"
            },
            "type": "array",
            "title": "Maintenance Rules"
          },
          "maintenance_logs": {
            "items": {
              "$ref": "#/components/schemas/SyncLogItem"
            },
            "type": "array",
            "title": "Maintenance Logs"
          }
        },
        "type": "object",
        "title": "SyncPushRequest",
        "description": "Batch mutation payload uploaded by the mobile client."
      },
      "SyncPushResponse": {
        "properties": {
          "synced_at": {
            "type": "string",
            "format": "date-time",
            "title": "Synced At"
          },
          "applied_assets": {
            "type": "integer",
            "title": "Applied Assets"
          },
          "applied_rules": {
            "type": "integer",
            "title": "Applied Rules"
          },
          "applied_logs": {
            "type": "integer",
            "title": "Applied Logs"
          }
        },
        "type": "object",
        "required": [
          "synced_at",
          "applied_assets",
          "applied_rules",
          "applied_logs"
        ],
        "title": "SyncPushResponse",
        "description": "Confirmation payload returned after processing sync mutations."
      },
      "SyncRuleItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "asset_id": {
            "type": "string",
            "format": "uuid",
            "title": "Asset Id"
          },
          "maintenance_type": {
            "type": "string",
            "title": "Maintenance Type"
          },
          "metric_type": {
            "$ref": "#/components/schemas/MetricType"
          },
          "interval_value": {
            "type": "number",
            "title": "Interval Value"
          },
          "metric_unit": {
            "type": "string",
            "title": "Metric Unit"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "is_deleted": {
            "type": "boolean",
            "title": "Is Deleted",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "asset_id",
          "maintenance_type",
          "metric_type",
          "interval_value",
          "metric_unit",
          "updated_at"
        ],
        "title": "SyncRuleItem",
        "description": "Maintenance rule mutation item submitted during push."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}